multimap would not compile in ruby 2.2
Per https://github.com/ruby/ruby/commit/8250aa2df0d6b4dcfa11dbad5307d28c2d5dd85f, RHASH is no longer provided by ruby.h.  Looks like multimap just needed an RHASH_IFNONE check anyways, so I've replaced it with that macro, which is available, and defined it to what it was prior to ruby 2.2 if it is not available.

Note that RHASH_IFNONE is deprecated, so this is just kicking the can
down the road a bit.